refactor: don't fetch suggestions for dismissed events
